Revert "Don's Sanity Check"
This reverts commit 7e36c371e971ec3490cb1e6b070ae2efff740be2.

+++ WebForm_DB_Createuser/WebForm_DB_Createuser/LiveCharts.aspx.cs
@@ -253,13 +253,13 @@ namespace WebForm_DB_Createuser
 
             //data format is [  {x1,y1}, {x2,y2}, ...]
             //date format: Date.UTC(2010, 1, 1, 12, 0, 0)
-            double avgTemp = 0;            
+            int avgTemp = 0;            
             foreach (DataRow dr in dt.Rows)
             {
                 string tempconverted = Convert.ToString(dr["tempValue"]);
                 string temp = tempconverted.Trim();
                 
-                avgTemp = Convert.ToDouble(temp);              
+                avgTemp = Convert.ToInt32(temp);              
 
             }
